Reorganized Manager code by service
