+++ src/CareTogether.Api/Controllers/VolunteersController.cs
@@ -1,4 +1,5 @@
 ﻿using CareTogether.Managers;
+using CareTogether.Managers.Approval;
 using CareTogether.Resources;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
